Modify index.js and styles.css to favour the program design
Set default container size to 16 and also declare the maximum size value.
Create a new header element inside the sketch pad above the container.
The header includes a display for current dimensions, and a reset button which uses JavaScript to allow change in the dimensions of the container.
Move the container creation inside the function generateContainer(), and also append the container inside the function at the end.
Tweak unique attributes for the generated cells inside the grid of the container for better correspondence to cartesian coordinates inside the DOM.
Reset button uses an event listener which promts user for new size for the container. On valid input argument use generateContainer() for the new size only after removing the old container node from the DOM.