Fix header display, and edit CSS styles
Replace size variable with default size constant for the initialisation.
Set id attribute for the header display, and remove textContent declaration before the header is appended to the sketch pad.
Inside generateContainer() function set the header display using the size provided as an argument when the function is called, and the query selector.