Add interactive cells feature, and minor changes in script and styles.
Use id attribute on individual cells to set unique coordinates.
Inside generateContainer() function attach event-listeners on each cell for color and shade.
Update conditions for reseting the container.
Use VIBGYOR as colors array, with first element being an empty string representing no color.
Disable context-menu open inside the sketch pad on right-click, as it will be used for interacting with the shade.
Write changeColor() and changeShade() functions which takes the cell on which the event is performed as an argument.
The function changeColor() loops through the colors in an order and sets the next color as the new cell background-color.
The function changeShade() loops between 0 and 0.9 by incrementing opacity by a factor of 10% on each interaction.